% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{copied_path}
\alias{copied_path}
\title{Format copied PC file location for universal use}
\usage{
copied_path()
}
\value{
After copying to clipboard, returns string formatted with standard file path formatting
}
\description{
Format copied PC file location for universal use
}
