% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Not In Operator}
\usage{
x \%ni\% table
}
\arguments{
\item{x}{The values to be looked fo}

\item{table}{The reference list to check whether items are in}
}
\value{
A logical vector indicating if a match was found
}
\description{
Not In Operator
}
\examples{
1 \%ni\% 1:10
c(10,20) \%in\% 20:24
}
