% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd2t.R
\name{hd2t}
\alias{hd2t}
\title{How to make herb and disease to target}
\usage{
hd2t(hb, disease, hbtype = "latin", diseasetype = "diseaseId")
}
\arguments{
\item{hb}{hb is the name of the drug}

\item{disease}{disease is the name of the disease}

\item{hbtype}{hbtype is the type of the hb}

\item{diseasetype}{diseasetype is the type of the disease}
}
\value{
targets and other information
}
\description{
How to make herb and disease to target
}
\examples{
hd2t(hb='mahuang',disease='C0009375',hbtype='pinyin',diseasetype='diseaseId')
hd2t(hb='Abutili Semen',disease='Colonic Neoplasms',hbtype='latin',diseasetype='diseaseName')
}
