% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tc_bcu}
\alias{tc_bcu}
\title{Cotización Dólar USA Billete}
\format{Un data-frame con cinco variables:
\describe{
\item{\code{Moneda}}{Moneda a cotizar}
\item{\code{Fecha}}{Fecha de cotización}
\item{\code{Venta}}{Precio venta}
\item{\code{Compra}}{Precio compra}
\item{\code{Arbitraje}}{Arbitraje}
}
Se tiene 4770 cotizaciones de dólar usa billete publicadas
entre el 03/01/2000 y el 28/12/2018.}
\usage{
tc_bcu
}
\description{
Cotización histórica Dólar USA Billete, desde el año 2000 a la fecha
}
\examples{
  tc_bcu
}
\keyword{datasets}
